# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MapVizit API - Map & Digital Business Card Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /points:
    get:
      operationId: listPoints
      tags: [points]
      summary: List all map points
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Points list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MapPoint"
    post:
      operationId: createPoint
      tags: [points]
      summary: Create a map point (sudo only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MapPointInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapPoint"
  /points/{id}:
    get:
      operationId: getPoint
      tags: [points]
      summary: Get point by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Point
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapPoint"
    patch:
      operationId: updatePoint
      tags: [points]
      summary: Update point (sudo or assigned admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MapPointUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapPoint"
    delete:
      operationId: deletePoint
      tags: [points]
      summary: Delete point (sudo only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /points/{id}/images:
    post:
      operationId: addPointImage
      tags: [points]
      summary: Add image to point
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PointImageInput"
      responses:
        "201":
          description: Image added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointImage"
  /points/{id}/images/{imageId}:
    delete:
      operationId: deletePointImage
      tags: [points]
      summary: Delete image from point
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /points/{id}/contacts:
    post:
      operationId: addPointContact
      tags: [points]
      summary: Add contact to point
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PointContactInput"
      responses:
        "201":
          description: Contact added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointContact"
  /points/{id}/contacts/{contactId}:
    delete:
      operationId: deletePointContact
      tags: [points]
      summary: Delete contact from point
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /points/{id}/assign-admin:
    post:
      operationId: assignAdminToPoint
      tags: [points]
      summary: Assign admin user to a point (sudo only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignAdminInput"
      responses:
        "200":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapPoint"
...
  /saved/{pointId}:
    post:
      operationId: savePoint
      tags: [saved]
      summary: Save a point
      parameters:
        - name: pointId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Saved
    delete:
      operationId: unsavePoint
      tags: [saved]
      summary: Unsave a point
      parameters:
        - name: pointId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unsaved

  # SETTINGS
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, icon, color]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        pointCount:
          type: integer
...
    PointContact:
      type: object
      required: [id, type, value]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [phone, telegram, instagram, website]
        value:
          type: string
        label:
          type: ["string", "null"]
    PointContactInput:
      type: object
      required: [type, value]
      properties:
        type:
          type: string
          enum: [phone, telegram, instagram, website]
        value:
          type: string
        label:
          type: ["string", "null"]
    PointImage:
      type: object
      required: [id, url]
      properties:
        id:
          type: integer
        url:
          type: string
        caption:
          type: ["string", "null"]
    PointImageInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
        caption:
          type: ["string", "null"]
    MapPoint:
      type: object
      required: [id, name, lat, lng, vizitkaCode, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        lat:
          type: number
        lng:
          type: number
        vizitkaCode:
          type: string
        categoryId:
          type: ["integer", "null"]
        category:
          $ref: "#/components/schemas/Category"
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/PointContact"
        images:
          type: array
          items:
            $ref: "#/components/schemas/PointImage"
        adminId:
          type: ["integer", "null"]
        isSaved:
          type: boolean
        createdAt:
          type: string
    MapPointInput:
      type: object
      required: [name, lat, lng]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        lat:
          type: number
        lng:
          type: number
        categoryId:
          type: ["integer", "null"]
    MapPointUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        lat:
          type: number
        lng:
          type: number
        categoryId:
          type: ["integer", "null"]
    AssignAdminInput:
      type: object
      required: [adminId]
      properties:
        adminId:
          type: ["integer", "null"]
...